refactor: split BudgetLoan.tsx into per-component files and useLoans hook
Extract CurrencyInput, LoanForm, RateHistoryPanel, and AmortizationTable
from the ~860-line BudgetLoan.tsx into individual files under
web/src/pages/budget/loan/. Move loans list/create/update/delete logic into
a new useLoans hook. Collapse the AmortizationTable's three-field cache
(loadedLoanId/loadedRows/loadedVersion) into a single cacheKey string.
Dedupe RateHistoryPanel's effective-rate formatting and drop the
meaningless > 0 ternary. Behavior is unchanged.

Bead: Hytte-ru94
