+++ web/src/pages/BudgetLoan.tsx
@@ -1,58 +1,12 @@
-import { useState, useEffect, useCallback, useRef } from 'react'
+import { useState } from 'react'
 import { Link } from 'react-router-dom'
-import { ArrowLeft, Home, Plus, Pencil, Trash2, ChevronDown, ChevronUp, X } from 'lucide-react'
+import { ArrowLeft, Home, Plus, Pencil, Trash2, ChevronDown, ChevronUp } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
-import type { TFunction } from 'i18next'
-
-interface Loan {
-  id: number
-  name: string
-  principal: number
-  current_balance: number
-  annual_rate: number
-  monthly_payment: number
-  start_date: string
-  first_payment_date: string
-  term_months: number
-  payment_day: number
-  property_value: number
-  property_name: string
-  notes: string
-  ltv_ratio?: number
-}
-
-interface AmortizationRow {
-  payment_num: number
-  date: string
-  payment: number
-  principal: number
-  interest: number
-  remaining_balance: number
-  rate: number
-}
-
-interface LoanRateChange {
-  id: number
-  loan_id: number
-  effective_date: string
-  annual_rate: number
-}
-
-interface AmortizationResponse {
-  loan: Loan
-  amortization: AmortizationRow[]
-  rate_changes: LoanRateChange[]
-  ltv_ratio: number
-  ltv_max: number
-}
-
-function localDateString(): string {
-  const d = new Date()
-  const y = d.getFullYear()
-  const m = String(d.getMonth() + 1).padStart(2, '0')
-  const day = String(d.getDate()).padStart(2, '0')
-  return `${y}-${m}-${day}`
-}
+import type { Loan } from './budget/loan/types'
+import { fmt, fmtPct, effectiveRate, localDateString } from './budget/loan/format'
+import { LoanForm } from './budget/loan/LoanForm'
+import { AmortizationTable } from './budget/loan/AmortizationTable'
+import { useLoans } from './budget/loan/useLoans'
 
 const EMPTY_LOAN: Omit<Loan, 'id'> = {
   name: '',
@@ -69,612 +23,36 @@ const EMPTY_LOAN: Omit<Loan, 'id'> = {
   notes: '',
 }
 
-function fmt(n: number): string {
-  return new Intl.NumberFormat('nb-NO', {
-    style: 'currency',
-    currency: 'NOK',
-    minimumFractionDigits: 0,
-    maximumFractionDigits: 0,
-  }).format(n)
-}
-
-function fmtPct(n: number): string {
-  return new Intl.NumberFormat(undefined, {
-    style: 'percent',
-    minimumFractionDigits: 1,
-    maximumFractionDigits: 2,
-  }).format(n)
-}
-
-/** Effective annual rate from nominal rate compounded monthly: (1 + r/12)^12 - 1 */
-function effectiveRate(nominalAnnual: number): number {
-  return Math.pow(1 + nominalAnnual / 12, 12) - 1
-}
-
-/** Number input that displays NOK-formatted value when not focused. */
-function CurrencyInput({ value, onChange, id, step, placeholder, min }: {
-  value: number
-  onChange: (v: number) => void
-  id: string
-  step?: string
-  placeholder?: string
-  min?: string
-}) {
-  const [focused, setFocused] = useState(false)
-  const [text, setText] = useState(String(value || ''))
-  const ref = useRef<HTMLInputElement>(null)
-
-  return (
-    <input
-      ref={ref}
-      id={id}
-      type={focused ? 'number' : 'text'}
-      min={min ?? '0'}
-      step={step ?? '1000'}
-      value={focused ? text : (value ? fmt(value) : '')}
-      placeholder={placeholder}
-      onFocus={() => {
-        setFocused(true)
-        setText(String(value || ''))
-      }}
-      onBlur={() => {
-        setFocused(false)
-        const n = Number(text)
-        if (!isNaN(n)) onChange(n)
-      }}
-      onChange={e => {
-        setText(e.target.value)
-        const n = Number(e.target.value)
-        if (!isNaN(n)) onChange(n)
-      }}
-      className="w-full bg-gray-700 border border-gray-600 rounded px-3 py-2 text-sm"
-    />
-  )
-}
-
-interface LoanFormProps {
-  initial: Omit<Loan, 'id'>
-  onSave: (loan: Omit<Loan, 'id'>) => Promise<void>
-  onCancel: () => void
-  saving: boolean
-  t: TFunction<'budget'>
-}
-
-function LoanForm({ initial, onSave, onCancel, saving, t }: LoanFormProps) {
-  const [form, setForm] = useState(initial)
-
-  function set<K extends keyof typeof form>(key: K, value: typeof form[K]) {
-    setForm(prev => ({ ...prev, [key]: value }))
-  }
-
-  function handleSubmit(e: React.FormEvent) {
-    e.preventDefault()
-    void onSave(form)
-  }
-
-  return (
-    <form onSubmit={handleSubmit} className="space-y-4">
-      <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-        <div>
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-name">
-            {t('loan.name')}
-          </label>
-          <input
-            id="loan-name"
-            type="text"
-            required
-            value={form.name}
-            onChange={e => set('name', e.target.value)}
-            className="w-full bg-gray-700 border border-gray-600 rounded px-3 py-2 text-sm"
-            placeholder={t('loan.namePlaceholder')}
-          />
-        </div>
-        <div>
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-start-date">
-            {t('loan.startDate')}
-          </label>
-          <input
-            id="loan-start-date"
-            type="date"
-            required
-            value={form.start_date}
-            onChange={e => set('start_date', e.target.value)}
-            className="w-full bg-gray-700 border border-gray-600 rounded px-3 py-2 text-sm"
-          />
-        </div>
-        <div>
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-first-payment">
-            {t('loan.firstPaymentDate')}
-          </label>
-          <input
-            id="loan-first-payment"
-            type="date"
-            value={form.first_payment_date}
-            onChange={e => set('first_payment_date', e.target.value)}
-            className="w-full bg-gray-700 border border-gray-600 rounded px-3 py-2 text-sm"
-          />
-          <p className="text-xs text-gray-500 mt-0.5">{t('loan.firstPaymentDateHint')}</p>
-        </div>
-        <div>
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-principal">
-            {t('loan.principal')}
-          </label>
-          <CurrencyInput id="loan-principal" value={form.principal} onChange={v => set('principal', v)} />
-        </div>
-        <div>
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-balance">
-            {t('loan.currentBalance')}
-          </label>
-          <CurrencyInput id="loan-balance" value={form.current_balance} onChange={v => set('current_balance', v)} />
-        </div>
-        <div>
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-rate">
-            {t('loan.annualRate')}
-          </label>
-          <div className="relative">
-            <input
-              id="loan-rate"
-              type="number"
-              min="0"
-              max="100"
-              step="0.01"
-              value={parseFloat((form.annual_rate * 100).toFixed(4))}
-              onChange={e => set('annual_rate', Number(e.target.value) / 100)}
-              className="w-full bg-gray-700 border border-gray-600 rounded px-3 py-2 pr-8 text-sm"
-              placeholder="4.80"
-            />
-            <span className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-400 text-sm pointer-events-none">%</span>
-          </div>
-          <p className="text-xs text-gray-500 mt-0.5">
-            {t('loan.effectiveRateHint', { pct: (effectiveRate(form.annual_rate) * 100).toFixed(2) })}
-          </p>
-        </div>
-        <div>
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-payment">
-            {t('loan.monthlyPayment')}
-          </label>
-          <CurrencyInput id="loan-payment" value={form.monthly_payment} onChange={v => set('monthly_payment', v)} step="100" placeholder="0" />
-          <p className="text-xs text-gray-500 mt-0.5">{t('loan.monthlyPaymentHint')}</p>
-        </div>
-        <div>
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-term">
-            {t('loan.termMonths')}
-          </label>
-          <input
-            id="loan-term"
-            type="number"
-            min="0"
-            step="12"
-            value={form.term_months}
-            onChange={e => set('term_months', Number(e.target.value))}
-            className="w-full bg-gray-700 border border-gray-600 rounded px-3 py-2 text-sm"
-          />
-        </div>
-        <div>
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-pay-day">
-            {t('loan.paymentDay')}
-          </label>
-          <input
-            id="loan-pay-day"
-            type="number"
-            min="1"
-            max="28"
-            value={form.payment_day}
-            onChange={e => set('payment_day', Number(e.target.value))}
-            className="w-full bg-gray-700 border border-gray-600 rounded px-3 py-2 text-sm"
-          />
-          <p className="text-xs text-gray-500 mt-0.5">{t('loan.paymentDayHint')}</p>
-        </div>
-        <div>
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-prop-value">
-            {t('loan.propertyValue')}
-          </label>
-          <CurrencyInput id="loan-prop-value" value={form.property_value} onChange={v => set('property_value', v)} step="10000" />
-        </div>
-        <div>
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-prop-name">
-            {t('loan.propertyName')}
-          </label>
-          <input
-            id="loan-prop-name"
-            type="text"
-            value={form.property_name}
-            onChange={e => set('property_name', e.target.value)}
-            className="w-full bg-gray-700 border border-gray-600 rounded px-3 py-2 text-sm"
-          />
-        </div>
-        <div className="sm:col-span-2">
-          <label className="block text-xs text-gray-400 mb-1" htmlFor="loan-notes">
-            {t('loan.notes')}
-          </label>
-          <textarea
-            id="loan-notes"
-            value={form.notes}
-            onChange={e => set('notes', e.target.value)}
-            rows={2}
-            className="w-full bg-gray-700 border border-gray-600 rounded px-3 py-2 text-sm resize-none"
-          />
-        </div>
-      </div>
-
-      <div className="flex gap-3 justify-end">
-        <button
-          type="button"
-          onClick={onCancel}
-          className="px-4 py-2 rounded-lg bg-gray-700 hover:bg-gray-600 text-sm transition-colors"
-        >
-          {t('quickAdd.cancel')}
-        </button>
-        <button
-          type="submit"
-          disabled={saving}
-          className="px-4 py-2 rounded-lg bg-blue-600 hover:bg-blue-500 text-sm font-medium transition-colors disabled:opacity-50"
-        >
-          {saving ? t('quickAdd.saving') : t('loan.save')}
-        </button>
-      </div>
-    </form>
-  )
-}
-
-/** Collapsible rate history panel with inline add form. */
-function RateHistoryPanel({ rateChanges, onAdd, onDelete, t }: {
-  rateChanges: LoanRateChange[]
-  onAdd: (date: string, rate: number) => Promise<void>
-  onDelete: (id: number) => Promise<void>
-  t: TFunction<'budget'>
-}) {
-  const [open, setOpen] = useState(false)
-  const [newDate, setNewDate] = useState('')
-  const [newRate, setNewRate] = useState('')
-  const [saving, setSaving] = useState(false)
-
-  async function handleAdd() {
-    if (!newDate || !newRate) return
-    setSaving(true)
-    await onAdd(newDate, Number(newRate) / 100)
-    setNewDate('')
-    setNewRate('')
-    setSaving(false)
-  }
-
-  return (
-    <div className="mb-4">
-      <button
-        onClick={() => setOpen(prev => !prev)}
-        className="flex items-center gap-1.5 text-sm text-gray-400 hover:text-gray-200 transition-colors"
-      >
-        {open ? <ChevronUp size={14} /> : <ChevronDown size={14} />}
-        {t('loan.rateHistory')}
-        {rateChanges.length > 0 && (
-          <span className="text-xs text-gray-500">({rateChanges.length})</span>
-        )}
-      </button>
-
-      {open && (
-        <div className="mt-2 space-y-2">
-          {rateChanges.map(rc => (
-            <div key={rc.id} className="flex items-center gap-3 text-sm">
-              <span className="text-gray-400">{rc.effective_date}</span>
-              <span className="font-medium">{(rc.annual_rate * 100).toFixed(2)}%</span>
-              <span className="text-xs text-gray-500">
-                ({t('loan.effectiveShort', { pct: (Math.pow(1 + rc.annual_rate / 12, 12) - 1) * 100 > 0 ? ((Math.pow(1 + rc.annual_rate / 12, 12) - 1) * 100).toFixed(2) : '0' })})
-              </span>
-              <button
-                onClick={() => void onDelete(rc.id)}
-                className="p-1 rounded hover:bg-gray-700 text-gray-500 hover:text-red-400 transition-colors"
-                aria-label={t('loan.delete')}
-              >
-                <X size={12} />
-              </button>
-            </div>
-          ))}
-
-          {rateChanges.length === 0 && (
-            <p className="text-xs text-gray-500">{t('loan.noRateChanges')}</p>
-          )}
-
-          {/* Add form */}
-          <div className="flex items-end gap-2 pt-1">
-            <div>
-              <label className="block text-xs text-gray-500 mb-0.5">{t('loan.date')}</label>
-              <input
-                type="date"
-                value={newDate}
-                onChange={e => setNewDate(e.target.value)}
-                className="bg-gray-700 border border-gray-600 rounded px-2 py-1 text-sm"
-              />
-            </div>
-            <div>
-              <label className="block text-xs text-gray-500 mb-0.5">{t('loan.annualRate')}</label>
-              <div className="relative">
-                <input
-                  type="number"
-                  min="0"
-                  max="100"
-                  step="0.01"
-                  value={newRate}
-                  onChange={e => setNewRate(e.target.value)}
-                  placeholder="4.80"
-                  className="bg-gray-700 border border-gray-600 rounded px-2 py-1 pr-7 text-sm w-24"
-                />
-                <span className="absolute right-2 top-1/2 -translate-y-1/2 text-gray-400 text-sm pointer-events-none">%</span>
-              </div>
-            </div>
-            <button
-              onClick={() => void handleAdd()}
-              disabled={saving || !newDate || !newRate}
-              className="px-3 py-1 bg-blue-600 hover:bg-blue-500 rounded text-sm font-medium transition-colors disabled:opacity-50"
-            >
-              <Plus size={14} />
-            </button>
-          </div>
-        </div>
-      )}
-    </div>
-  )
-}
-
-const INITIAL_AMORTIZATION_ROWS = 24
-const FULL_AMORTIZATION_ROWS = 360
-
-interface AmortizationTableProps {
-  loanId: number
-  version: number
-  t: TFunction<'budget'>
-}
-
-function AmortizationTable({ loanId, version, t }: AmortizationTableProps) {
-  const [data, setData] = useState<AmortizationResponse | null>(null)
-  const [loading, setLoading] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-  const [showAll, setShowAll] = useState(false)
-  const [showPast, setShowPast] = useState(false)
-  const [loadedLoanId, setLoadedLoanId] = useState<number | null>(null)
-  const [loadedRows, setLoadedRows] = useState(0)
-  const [loadedVersion, setLoadedVersion] = useState(0)
-
-  useEffect(() => {
-    const requestedRows = showAll ? FULL_AMORTIZATION_ROWS : INITIAL_AMORTIZATION_ROWS
-
-    if (loadedLoanId === loanId && loadedRows >= requestedRows && loadedVersion === version) {
-      return
-    }
-
-    const controller = new AbortController()
-    // eslint-disable-next-line react-hooks/set-state-in-effect -- async data fetch; AbortController prevents stale updates on unmount
-    setLoading(true)
-    setError(null)
-    fetch(`/api/budget/loans/${loanId}/amortization?rows=${requestedRows}`, { credentials: 'include', signal: controller.signal })
-      .then(r => {
-        if (!r.ok) throw new Error('fetch failed')
-        return r.json() as Promise<AmortizationResponse>
-      })
-      .then(response => {
-        setData(response)
-        setLoadedLoanId(loanId)
-        setLoadedRows(requestedRows)
-        setLoadedVersion(version)
-      })
-      .catch(err => {
-        if (err instanceof Error && err.name === 'AbortError') return
-        setError(t('loan.errors.loadFailed'))
-      })
-      .finally(() => setLoading(false))
-    return () => controller.abort()
-  }, [loanId, loadedLoanId, loadedRows, loadedVersion, version, showAll, t])
-
-  if (loading) return <p className="text-gray-400 text-sm py-4">{t('loading')}</p>
-  if (error) return <p className="text-red-400 text-sm py-4">{error}</p>
-  if (!data || data.amortization.length === 0) {
-    return <p className="text-gray-500 text-sm py-4">{t('loan.noAmortization')}</p>
-  }
-
-  const ltvPct = data.ltv_ratio
-  const ltvOk = ltvPct <= data.ltv_max
-  const allRows = data.amortization
-  const rateChanges = data.rate_changes ?? []
-  const today = localDateString()
-  const pastRows = allRows.filter(r => r.date <= today)
-  const futureRows = allRows.filter(r => r.date > today)
-  const visibleRows = showPast ? allRows : futureRows
-
-  async function addRateChange(effectiveDate: string, annualRate: number) {
-    try {
-      const r = await fetch(`/api/budget/loans/${loanId}/rates`, {
-        method: 'POST',
-        credentials: 'include',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ effective_date: effectiveDate, annual_rate: annualRate }),
-      })
-      if (!r.ok) throw new Error('create failed')
-      // Force reload amortization
-      setLoadedLoanId(null)
-    } catch {
-      setError(t('loan.errors.saveFailed'))
-    }
-  }
-
-  async function deleteRateChange(rcId: number) {
-    try {
-      const r = await fetch(`/api/budget/loans/${loanId}/rates/${rcId}`, {
-        method: 'DELETE',
-        credentials: 'include',
-      })
-      if (!r.ok) throw new Error('delete failed')
-      setLoadedLoanId(null)
-    } catch {
-      setError(t('loan.errors.deleteFailed'))
-    }
-  }
-
-  return (
-    <div>
-      {/* LTV display */}
-      {data.loan.property_value > 0 && (
-        <div className="mb-4 flex items-center gap-4">
-          <span className="text-sm text-gray-400">{t('loan.ltv')}</span>
-          <div className="flex-1 bg-gray-700 rounded-full h-2 max-w-xs">
-            <div
-              className={`h-2 rounded-full transition-all ${ltvOk ? 'bg-green-500' : 'bg-red-500'}`}
-              style={{ width: `${Math.min(ltvPct / data.ltv_max, 1) * 100}%` }}
-            />
-          </div>
-          <span className={`text-sm font-medium ${ltvOk ? 'text-green-400' : 'text-red-400'}`}>
-            {fmtPct(ltvPct)}
-          </span>
-          <span className="text-xs text-gray-500">{t('loan.ltvMax', { pct: fmtPct(data.ltv_max) })}</span>
-        </div>
-      )}
-
-      {/* Rate history */}
-      <RateHistoryPanel
-        rateChanges={rateChanges}
-        onAdd={addRateChange}
-        onDelete={deleteRateChange}
-        t={t}
-      />
-
-      {/* Past payments toggle */}
-      {pastRows.length > 0 && (
-        <button
-          onClick={() => setShowPast(prev => !prev)}
-          className="mb-3 flex items-center gap-1.5 text-sm text-gray-400 hover:text-gray-200 transition-colors"
-        >
-          {showPast ? <ChevronUp size={14} /> : <ChevronDown size={14} />}
-          {showPast
-            ? t('loan.hidePastPayments')
-            : t('loan.showPastPayments', { count: pastRows.length })}
-        </button>
-      )}
-
-      <div className="overflow-x-auto">
-        <table className="w-full text-sm">
-          <thead>
-            <tr className="text-left text-gray-400 border-b border-gray-700">
-              <th className="pb-2 pr-3">#</th>
-              <th className="pb-2 pr-3">{t('loan.date')}</th>
-              <th className="pb-2 pr-3 text-right">{t('loan.payment')}</th>
-              <th className="pb-2 pr-3 text-right">{t('loan.principalPart')}</th>
-              <th className="pb-2 pr-3 text-right">{t('loan.interestPart')}</th>
-              <th className="pb-2 pr-3 text-right">{t('loan.remainingBalance')}</th>
-              <th className="pb-2 text-right">{t('loan.rate')}</th>
-            </tr>
-          </thead>
-          <tbody className="divide-y divide-gray-800">
-            {visibleRows.map(row => (
-              <tr key={row.payment_num} className={`hover:bg-gray-700/30 ${row.date <= today ? 'text-gray-500' : 'text-gray-300'}`}>
-                <td className="py-1.5 pr-3 text-gray-500">{row.payment_num}</td>
-                <td className="py-1.5 pr-3">{row.date}</td>
-                <td className="py-1.5 pr-3 text-right">{fmt(row.payment)}</td>
-                <td className={`py-1.5 pr-3 text-right ${row.date <= today ? 'text-blue-400/50' : 'text-blue-400'}`}>{fmt(row.principal)}</td>
-                <td className={`py-1.5 pr-3 text-right ${row.date <= today ? 'text-red-400/50' : 'text-red-400'}`}>{fmt(row.interest)}</td>
-                <td className="py-1.5 pr-3 text-right">{fmt(row.remaining_balance)}</td>
-                <td className="py-1.5 text-right text-gray-400">{(row.rate * 100).toFixed(2)}%</td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-
-      {(allRows.length >= INITIAL_AMORTIZATION_ROWS || showAll) && (
-        <button
-          onClick={() => setShowAll(prev => !prev)}
-          className="mt-3 flex items-center gap-1 text-sm text-blue-400 hover:text-blue-300"
-        >
-          {showAll ? (
-            <><ChevronUp size={14} />{t('loan.showLess')}</>
-          ) : (
-            <><ChevronDown size={14} />{t('loan.showAll', { count: allRows.length })}</>
-          )}
-        </button>
-      )}
-    </div>
-  )
-}
-
 export default function BudgetLoan() {
   const { t } = useTranslation('budget')
-  const [loans, setLoans] = useState<Loan[]>([])
-  const [loading, setLoading] = useState(true)
-  const [error, setError] = useState<string | null>(null)
+  const { loans, loading, error, createLoan, updateLoan, deleteLoan } = useLoans()
   const [showForm, setShowForm] = useState(false)
   const [editingLoan, setEditingLoan] = useState<Loan | null>(null)
   const [saving, setSaving] = useState(false)
   const [expandedAmortization, setExpandedAmortization] = useState<number | null>(null)
   const [loanVersion, setLoanVersion] = useState(0)
 
-  const fetchLoans = useCallback(async () => {
-    setError(null)
-    try {
-      const r = await fetch('/api/budget/loans', { credentials: 'include' })
-      if (!r.ok) throw new Error('fetch failed')
-      const d = await r.json() as { loans: Loan[] }
-      setLoans(d.loans)
-    } catch {
-      setError(t('loan.errors.loadFailed'))
-    } finally {
-      setLoading(false)
-    }
-  }, [t])
-
-  useEffect(() => {
-    // eslint-disable-next-line react-hooks/set-state-in-effect -- async data fetch; AbortController prevents stale updates on unmount
-    void fetchLoans()
-  }, [fetchLoans])
-
   async function handleCreate(form: Omit<Loan, 'id'>) {
     setSaving(true)
-    try {
-      const r = await fetch('/api/budget/loans', {
-        method: 'POST',
-        credentials: 'include',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(form),
-      })
-      if (!r.ok) throw new Error('create failed')
-      setShowForm(false)
-      await fetchLoans()
-    } catch {
-      setError(t('loan.errors.saveFailed'))
-    } finally {
-      setSaving(false)
-    }
+    const ok = await createLoan(form)
+    if (ok) setShowForm(false)
+    setSaving(false)
   }
 
   async function handleUpdate(form: Omit<Loan, 'id'>) {
     if (!editingLoan) return
     setSaving(true)
-    try {
-      const r = await fetch(`/api/budget/loans/${editingLoan.id}`, {
-        method: 'PUT',
-        credentials: 'include',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(form),
-      })
-      if (!r.ok) throw new Error('update failed')
+    const ok = await updateLoan(editingLoan.id, form)
+    if (ok) {
       setEditingLoan(null)
       setLoanVersion(v => v + 1)
-      await fetchLoans()
-    } catch {
-      setError(t('loan.errors.saveFailed'))
-    } finally {
-      setSaving(false)
     }
+    setSaving(false)
   }
 
   async function handleDelete(id: number) {
     if (!confirm(t('loan.confirmDelete'))) return
-    try {
-      const r = await fetch(`/api/budget/loans/${id}`, {
-        method: 'DELETE',
-        credentials: 'include',
-      })
-      if (!r.ok) throw new Error('delete failed')
-      await fetchLoans()
-    } catch {
-      setError(t('loan.errors.deleteFailed'))
-    }
+    await deleteLoan(id)
   }
 
   return (
