+++ web/src/pages/budget/loan/format.ts
@@ -1,0 +1,29 @@
+export function fmt(n: number): string {
+  return new Intl.NumberFormat('nb-NO', {
+    style: 'currency',
+    currency: 'NOK',
+    minimumFractionDigits: 0,
+    maximumFractionDigits: 0,
+  }).format(n)
+}
+
+export function fmtPct(n: number): string {
+  return new Intl.NumberFormat(undefined, {
+    style: 'percent',
+    minimumFractionDigits: 1,
+    maximumFractionDigits: 2,
+  }).format(n)
+}
+
+/** Effective annual rate from nominal rate compounded monthly: (1 + r/12)^12 - 1 */
+export function effectiveRate(nominalAnnual: number): number {
+  return Math.pow(1 + nominalAnnual / 12, 12) - 1
+}
+
+export function localDateString(): string {
+  const d = new Date()
+  const y = d.getFullYear()
+  const m = String(d.getMonth() + 1).padStart(2, '0')
+  const day = String(d.getDate()).padStart(2, '0')
+  return `${y}-${m}-${day}`
+}
